feat: add handleAuthError hook and session expiry modal

+++ hooks/CartPageHook/useAddToCart.ts
@@ -8,25 +8,28 @@ import { DeleteItemFromCart } from '../../services/api/cart-apis/remove-item-api
 import { CONSTANTS } from '../../services/config/app-config';
 import { get_access_token, storeToken } from '../../store/slices/auth/token-login-slice';
 import { addCartList, addItemToCart, clearCart, removeItemFromCart } from '../../store/slices/cart-slices/cart-local-slice';
+import useAuthErrorHandler from '../AuthHooks/handleAuthError';
 
 const useAddToCartHook = () => {
   const dispatch = useDispatch();
   const tokenFromStore: any = useSelector(get_access_token);
   const getPartyName = localStorage.getItem('party_name');
   const { SUMMIT_APP_CONFIG }: any = CONSTANTS;
+  const handleAuthError = useAuthErrorHandler();
   const extractProductCodes = (data: any[]) => {
     return data?.flatMap((category) => category.orders.map((order: any) => order.item_code));
   };
   const getCartList = async (setCartListingItems: any) => {
     try {
       let cartListingData: any = await fetchCartListingAPI(SUMMIT_APP_CONFIG, tokenFromStore.token);
-      if (cartListingData.data.message.msg === 'success') {
+      if (cartListingData?.status === 200 && cartListingData.data.message.msg === 'success') {
         setCartListingItems(cartListingData?.data?.message?.data);
         let cartData = extractProductCodes(cartListingData?.data?.message?.data?.categories);
         let quotationId = cartListingData?.data?.message?.data?.name;
         dispatch(addCartList({ cartData, quotationId }));
       } else {
         setCartListingItems({});
+        handleAuthError(cartListingData);
       }
       return cartListingData;
     } catch (error) {
@@ -44,26 +47,29 @@ const useAddToCartHook = () => {
       }
     } else {
       toast.error(postDataInCart?.data?.message?.error);
+      handleAuthError(postDataInCart);
     }
   };
   const placeOrderAPIFunc = async (params: any, setCartListingItems: any) => {
     const placeOrder = await postPlaceOrderAPI(SUMMIT_APP_CONFIG, params, tokenFromStore?.token);
-    if (placeOrder?.data?.message?.msg === 'success') {
+    if (placeOrder?.status === 200 && placeOrder?.data?.message?.msg === 'success') {
       dispatch(clearCart());
       toast.success('Order placed successfully!');
       setCartListingItems({});
     } else {
       toast.error('Failed to place order.');
+      handleAuthError(placeOrder);
     }
   };
   const RemoveItemCartAPIFunc = async (params: any, setCartListingItems: any) => {
     const removeCartfunc = await DeleteItemFromCart(SUMMIT_APP_CONFIG, params, tokenFromStore?.token);
-    if (removeCartfunc?.data?.message?.msg === 'success') {
+    if (removeCartfunc?.status === 200 && removeCartfunc?.data?.message?.msg === 'success') {
       dispatch(removeItemFromCart(params?.item_code));
       toast.success('Product removed from cart successfully!');
       getCartList(setCartListingItems);
     } else {
       toast.error('Failed to remove product from cart');
+      handleAuthError(removeCartfunc);
     }
   };
   const cLearCartAPIFunc = async (quotation_id: any, setCartListingItems: any, setClearCartLoader: any) => {
@@ -76,6 +82,7 @@ const useAddToCartHook = () => {
         setCartListingItems({});
         toast.success('Cart cleared successfully!');
       } else {
+        handleAuthError(clearCartfunc);
         toast.error('Failed to clear cart.');
       }
     } catch (error) {
